[Dev-02] Added solution to the first problem.
InsuranceRequirements has methods for evaluating travel insurance criteria.
Passenger can be used to create passenger for trips, and evaluate them if they match the above criteria.
TravelFor is an enumeration listing the possible purposes for a trip.
Test is used for generating and evaluating passengers and testing the results of these evaluations.